Trying again

+++ src/main/java/com/ac/umkc/spark/SparkDriver.java
@@ -667,7 +667,7 @@ public class SparkDriver implements Serializable {
       TweetsDayData data = new TweetsDayData();
       data.setUserType(row.getString(0));
       data.setShortDate(row.getString(1));
-      data.setCount((int)row.getLong(2));
+      data.setCount(row.getInt(2));
       data.setYear((int)row.getLong(3));
       data.setMonth((int)row.getLong(4));
       data.setDay((int)row.getLong(5));
